update integration tests
