+++ tests/integration/dedicated_alb/test_dedicated_alb_deprovisioning.py
@@ -4,6 +4,9 @@ from broker.extensions import db
 from broker.models import Operation, DedicatedALBServiceInstance
 from tests.lib import factories
 from tests.lib.client import check_last_operation_description
+from tests.lib.alb.deprovision import (
+    subtest_deprovision_removes_certificate_from_iam,
+)
 
 
 @pytest.fixture
@@ -155,31 +158,6 @@ def subtest_deprovision_removes_cert_from_alb(tasks, service_instance, alb):
     )
     tasks.run_queued_tasks_and_enqueue_dependents()
     alb.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_certificate_from_iam(
-    tasks, service_instance, iam_govcloud
-):
-    service_instance = db.session.get(DedicatedALBServiceInstance, "1234")
-    iam_govcloud.expects_delete_server_certificate(
-        service_instance.new_certificate.iam_server_certificate_name
-    )
-    iam_govcloud.expects_delete_server_certificate(
-        service_instance.current_certificate.iam_server_certificate_name
-    )
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    iam_govcloud.assert_no_pending_responses()
-
-
-def subtest_deprovision_removes_certificate_from_iam_when_missing(
-    tasks, service_instance, iam_govcloud
-):
-    service_instance = db.session.get(DedicatedALBServiceInstance, "1234")
-    iam_govcloud.expects_delete_server_certificate_returning_no_such_entity(
-        name=service_instance.iam_server_certificate_name
-    )
-    tasks.run_queued_tasks_and_enqueue_dependents()
-    iam_govcloud.assert_no_pending_responses()
 
 
 def subtest_deprovision_marks_operation_as_succeeded(tasks):
